safari-600.7-branch/LayoutTests/fast/dom/css-RGBValue-expected.txt

This test ensures that objects RGBColor objects return the correct values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(element, '').getPropertyCSSValue('color').getRGBColorValue().red.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 10
PASS window.getComputedStyle(element, '').getPropertyCSSValue('color').getRGBColorValue().green.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 20
PASS window.getComputedStyle(element, '').getPropertyCSSValue('color').getRGBColorValue().blue.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 30
PASS successfullyParsed is true

TEST COMPLETE
